Inverted annotation map
Before, `MGLMapView` held an `NSDictionary` mapping `MGLAnnotation`s to context dictionaries that contained the annotation ID and accessibility element. But it turned out to be more common to need the annotation or accessibility element given an ID. This change inverts the map and turns it into an `std::map`, which allows us to use primitive keys as well as plenty of C++ algorithmic goodness.

Also pulled annotation touch region tolerances into constants for use by the accessibility code.