Refactored MGLAnnotationContext
`MGLAnnotationContext` is a C++ class rather than a map. Removed assertions that were important when it was a map (because merely looking up a value by key would create an entry for that key). Fixed an issue where the “return to map” accessibility element lacked a label. Always update the title and subtitle of an annotation accessibility element, in case the title or subtitle changes from one showing to the next.